\name{continuousScheme}
\Rdversion{1.1}
\alias{continuousScheme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function continuousScheme
}
\description{
One line description of continuousScheme
}
\usage{
continuousScheme(valueColours)
}
\arguments{
  \item{valueColours}{argument  valueColours}
}

\details{
The details of continuousScheme
}

\value{
The value of continuousScheme
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

