\name{f3Scheme}
\Rdversion{1.1}
\alias{f3Scheme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function f3Scheme
}
\description{
One line description of f3Scheme
}
\usage{
f3Scheme(d3, f3, schemeText, equalScale = FALSE)
}
\arguments{
  \item{d3}{argument  d3}
  \item{f3}{argument  f3}
  \item{schemeText}{argument  schemeText}
  \item{equalScale}{argument  equalScale}
}

\details{
The details of f3Scheme
}

\value{
The value of f3Scheme
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

