\name{hsvScheme}
\Rdversion{1.1}
\alias{hsvScheme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function hsvScheme
}
\description{
One line description of hsvScheme
}
\usage{
hsvScheme(d3, equalScale = FALSE)
}
\arguments{
  \item{d3}{argument  d3}
  \item{equalScale}{argument  equalScale}
}

\details{
The details of hsvScheme
}

\value{
The value of hsvScheme
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

