\name{nearestScheme}
\Rdversion{1.1}
\alias{nearestScheme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function nearestScheme
}
\description{
One line description of nearestScheme
}
\usage{
nearestScheme(valueColours)
}
\arguments{
  \item{valueColours}{argument  valueColours}
}

\details{
The details of nearestScheme
}

\value{
The value of nearestScheme
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

