\name{print.colourScheme}
\Rdversion{1.1}
\alias{print.colourScheme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function print.colourScheme
}
\description{
One line description of print.colourScheme
}
\usage{
print.colourScheme(x, ...)
}
\arguments{
  \item{x}{argument  x}
  \item{\dots}{argument  \dots}
}

\details{
The details of print.colourScheme
}

\value{
The value of print.colourScheme
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

