\name{rampInterpolate}
\Rdversion{1.1}
\alias{rampInterpolate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function rampInterpolate
}
\description{
One line description of rampInterpolate
}
\usage{
rampInterpolate(limits, ramp, ...)
}
\arguments{
  \item{limits}{argument  limits}
  \item{ramp}{argument  ramp}
  \item{\dots}{argument  \dots}
}

\details{
The details of rampInterpolate
}

\value{
The value of rampInterpolate
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

