\name{rgbScheme}
\Rdversion{1.1}
\alias{rgbScheme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function rgbScheme
}
\description{
One line description of rgbScheme
}
\usage{
rgbScheme(d3, equalScale = FALSE)
}
\arguments{
  \item{d3}{argument  d3}
  \item{equalScale}{argument  equalScale}
}

\details{
The details of rgbScheme
}

\value{
The value of rgbScheme
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

