\name{schemeData.f3scheme}
\Rdversion{1.1}
\alias{schemeData.f3scheme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function schemeData.f3scheme
}
\description{
One line description of schemeData.f3scheme
}
\usage{
schemeData.f3scheme(object)
}
\arguments{
  \item{object}{argument  object}
}

\details{
The details of schemeData.f3scheme
}

\value{
The value of schemeData.f3scheme
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

