\name{schemeData.factorScheme}
\Rdversion{1.1}
\alias{schemeData.factorScheme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function schemeData.factorScheme
}
\description{
One line description of schemeData.factorScheme
}
\usage{
schemeData.factorScheme(object)
}
\arguments{
  \item{object}{argument  object}
}

\details{
The details of schemeData.factorScheme
}

\value{
The value of schemeData.factorScheme
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

