\name{schemeData.multiRamp}
\Rdversion{1.1}
\alias{schemeData.multiRamp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function schemeData.multiRamp
}
\description{
One line description of schemeData.multiRamp
}
\usage{
schemeData.multiRamp(object)
}
\arguments{
  \item{object}{argument  object}
}

\details{
The details of schemeData.multiRamp
}

\value{
The value of schemeData.multiRamp
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

