\name{schemeData.nearestScheme}
\Rdversion{1.1}
\alias{schemeData.nearestScheme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function schemeData.nearestScheme
}
\description{
One line description of schemeData.nearestScheme
}
\usage{
schemeData.nearestScheme(object)
}
\arguments{
  \item{object}{argument  object}
}

\details{
The details of schemeData.nearestScheme
}

\value{
The value of schemeData.nearestScheme
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

