\name{schemeData.rampInterpolate}
\Rdversion{1.1}
\alias{schemeData.rampInterpolate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function schemeData.rampInterpolate
}
\description{
One line description of schemeData.rampInterpolate
}
\usage{
schemeData.rampInterpolate(object)
}
\arguments{
  \item{object}{argument  object}
}

\details{
The details of schemeData.rampInterpolate
}

\value{
The value of schemeData.rampInterpolate
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

