\name{schemeData}
\Rdversion{1.1}
\alias{schemeData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function schemeData
}
\description{
One line description of schemeData
}
\usage{
schemeData(object)
}
\arguments{
  \item{object}{argument  object}
}

\details{
The details of schemeData
}

\value{
The value of schemeData
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

