\name{factorScheme}
\Rdversion{1.1}
\alias{factorScheme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function factorScheme
}
\description{
One line description of factorScheme
}
\usage{
factorScheme(fac, colours)
}
\arguments{
  \item{fac}{argument  fac}
  \item{colours}{argument  colours}
}

\details{
The details of factorScheme
}

\value{
The value of factorScheme
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

