\name{multiRamp}
\Rdversion{1.1}
\alias{multiRamp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Function multiRamp
}
\description{
One line description of multiRamp
}
\usage{
multiRamp(ranges, rampSpec, ...)
}
\arguments{
  \item{ranges}{argument  ranges}
  \item{rampSpec}{argument  rampSpec}
  \item{\dots}{argument  \dots}
}

\details{
The details of multiRamp
}

\value{
The value of multiRamp
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

