\name{schemeData.factorScheme}
\Rdversion{1.1}
\alias{schemeData.factorScheme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Return scheme data for a factor scheme
}
\description{
Return scheme data for a factor scheme
}
\usage{
\method{schemeData}{factorScheme}(object)
}
\arguments{
  \item{object}{a factorScheme object}
}

\value{
Returns a list of the scheme data for the colour scheme.
}

\author{
Barry Rowlingson <b.rowlingson@lancaster.ac.uk>
}



\keyword{color}
  

